requirements.txt
python-dotenv
sqlmodel
fastapi[standard]
pyjwt
passlib
bcrypt